requirements.txt
python-dotenv
streamlit==1.24.0
google-generativeai
